Introduction of Offline Synchronizer and Plotter Update
Offline Synchronizer:

This code can be used to synchronize the capture NAV and FAC data offline. It simulates the real-time environments using the timestamp and replicates the online synchronization results. This can be useful in creating possible scenarios manually and testing the behavior of the synchronizer. Please note that the synchronization function here remains the same as the online one it feeds the data to the functions in a similar manner as in the real-time scenario.

Plotter Updates:

The Plotter can now take lagged message with any number of FACs (based on the ones we have) and plot the graphs, without requiring any changes in the code. It is now fully automated.